alias: move alias parsing to alias module
Also, allow specifying %{CONFIG:key} values in alias files to reference
the related setting from the config file(s). For example, if the option
"user: bob" is set in the config file, then %{CONFIG:user} will be
expanded to "bob" if used within the alias file, assuming that the
connection sections match.