service._reqs: req_cmd(): initial support for passing object args to writable funcs
Methods such as modify, create, and attach should support passing
objects to them allowing for the translation to requests to be made
without forcing the user to manually do so.

This can allow for things such as creating new bugs locally and then
pushing the objects later to create them on the related service.