args.jira: add initial --project support for global projects
For connections configured to use a global jira instance, searches and
other subcmds will operate across all projects hosted on the instance.
This option restricts the subcmd to run across a specific project on the
jira service.

Useful in the case where you want to use a configured connection for the
entire jira server instead of one per project, but sometimes want to
restriction subcmd actions to a specific project.