service: be more lenient in allowing str values for various API calls
If we're going to let consumers directly call API commands, it'll help
to convert args on the fly similar to how they're converted by argparser
for input commands.

In fact, long term it might be reasonable to extract the type
checking/conversion method from the argparse side to use for direct API
calls or merge the two somehow so cli tool usage doesn't check things
twice but we still get type conversion for API usage.